requirements.txt
streamlit
python-dotenv
agno
requests
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
